CurrencyApp/app/src/test/java/com/example/currencyapp/currencyList/CurrencyListViewModelTest.kt
package com.example.currencyapp.currencyList

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import kotlinx.coroutines.Dispatchers
import kotlinx.coroutines.Job
import kotlinx.coroutines.newSingleThreadContext
import kotlinx.coroutines.runBlocking
import kotlinx.coroutines.test.resetMain
import kotlinx.coroutines.test.setMain
import org.junit.After
import org.junit.Assert.assertEquals
import org.junit.Assert.assertNotNull
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith
import org.junit.runners.JUnit4


@RunWith(JUnit4::class)
class CurrencyListViewModelTest {

    @get:Rule
    val instantTaskExecutorRule = InstantTaskExecutorRule()

    private val mainThreadSurrogate = newSingleThreadContext("Main thread")
    private var viewModelJob  = Job()
    lateinit var viewModel: CurrencyListViewModel


    @Before
    fun setUp() {
        Dispatchers.setMain(mainThreadSurrogate)
        viewModel = CurrencyListViewModel()
    }

    @After
    fun tearDown() {
        Dispatchers.resetMain() // reset main dispatcher to the original Main dispatcher
        mainThreadSurrogate.close()
    }


   @Test
    fun getList() =runBlocking (Dispatchers.Main+viewModelJob) {

                viewModel.getList()
                assertNotNull(viewModel.currencyRatesList.value!!)
                assertEquals("CAD",viewModel.currencyRatesList.value!!.elementAt(0))


        }



}